// embedded-components/src/api/generated/smbdo.schemas.ts
export type YearOfFormation = string;

/**
 * Whether the party has a website. False if opted out.
 */
export type WebsiteAvailable = boolean;

/**
 * Name/URL of the website
 * @minLength 1
 * @maxLength 500
 */
export type Website = string;

/**
 * The ID type
 */
export type OrganizationIdentityDtoIdType =
  (typeof OrganizationIdentityDtoIdType)[keyof typeof OrganizationIdentityDtoIdType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrganizationIdentityDtoIdType = {
  EIN: 'EIN',
  BUSINESS_REGISTRATION_ID: 'BUSINESS_REGISTRATION_ID',
  BUSINESS_NUMBER: 'BUSINESS_NUMBER',
  BUSINESS_REGISTRATION_NUMBER: 'BUSINESS_REGISTRATION_NUMBER',
} as const;

/**
 * The tax ID of the organization.
 */
export interface OrganizationIdentityDto {
  /** Description of the ID. */
  description?: string;
  expiryDate?: string;
  /** The ID type */
  idType: OrganizationIdentityDtoIdType;
  /**
   * Issuing authority
   * @minLength 1
   * @maxLength 500
   */
  issuer: string;
  /**
   * ID value
   * @minLength 1
   * @maxLength 100
   */
  value: string;
}

export type OrganizationType =
  (typeof OrganizationType)[keyof typeof OrganizationType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrganizationType = {
  LIMITED_LIABILITY_PARTNERSHIP: 'LIMITED_LIABILITY_PARTNERSHIP',
  LIMITED_PARTNERSHIP: 'LIMITED_PARTNERSHIP',
  GENERAL_PARTNERSHIP: 'GENERAL_PARTNERSHIP',
  LIMITED_LIABILITY_COMPANY: 'LIMITED_LIABILITY_COMPANY',
  C_CORPORATION: 'C_CORPORATION',
  S_CORPORATION: 'S_CORPORATION',
  PARTNERSHIP: 'PARTNERSHIP',
  PUBLICLY_TRADED_COMPANY: 'PUBLICLY_TRADED_COMPANY',
  NON_PROFIT_CORPORATION: 'NON_PROFIT_CORPORATION',
  GOVERNMENT_ENTITY: 'GOVERNMENT_ENTITY',
  SOLE_PROPRIETORSHIP: 'SOLE_PROPRIETORSHIP',
  UNINCORPORATED_ASSOCIATION: 'UNINCORPORATED_ASSOCIATION',
} as const;

/**
 * The organization?s description.

 * @minLength 0
 */
export type OrganizationDescription = string;

/**
 * The organization?s legal name. It is the official name of the person or entity that owns a company. Must be the name used on the legal party's government forms and business paperwork

 * @minLength 1
 */
export type OrganizationName = string;

/**
 * The merchant category code describing the merchant's industry.
 * @minLength 4
 * @maxLength 4
 */
export type MerchantCategoryCode = string;

export type OrganizationIndustryCodeType =
  (typeof OrganizationIndustryCodeType)[keyof typeof OrganizationIndustryCodeType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrganizationIndustryCodeType = {
  NAICS: 'NAICS',
} as const;

export interface OrganizationIndustry {
  /**
   * Value of the industry code such as a 9 digit NAICS code
   * @minLength 1
   * @maxLength 24
   */
  code?: string;
  codeType?: OrganizationIndustryCodeType;
}

/**
 * The industry type of the business connected to the client. You can use the Reference Data resource to get a list of acceptable values.

 * @minLength 0
 */
export type OrganizationIndustryType = string;

/**
 * The industry category of the business connected to the client. For example, Accommodation and Food Services. You can use the Reference Data resource to get a list of acceptable values.

 * @minLength 0
 */
export type OrganizationIndustryCategory = string;

/**
 * Entities in ownership means that one or more businesses own part of the business connected to the client. Always required for a Privately Owned Business.

 */
export type EntitiesInOwnership = boolean;

/**
 * @minLength 1
 * @maxLength 100
 */
export type DbaName = string;

export type IndividualDetailsRequired = IndividualDetails;

export type SoleOwner = boolean;
// ...
export interface SocialMedia {
  /** The platform of the social media profile.
| Platform | Description |
| -- | -- |
| FACEBOOK | Facebook, the social media profile at facebook.com. |
| INSTAGRAM | Instagram, the social media profile at instagram.com. |
| X | X (formerly known as Twitter), the social media profile at x.com. |
 */
  profilePlatform: SocialMediaProfilePlatform;
  /**
   * The username of the social media profile.
   * @minLength 1
   * @maxLength 50
   */
  username: string;
}

/**
 * The entity's social media profiles.
 * @minItems 1
 * @maxItems 3
 */
export type SocialMediaList = SocialMedia[];

/**
 * Nature of ownership e.g. Direct|Indirect

 * @minLength 0
 */
export type NatureOfOwnership = string;

export type PhoneSmbdoPhoneType =
  (typeof PhoneSmbdoPhoneType)[keyof typeof PhoneSmbdoPhoneType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PhoneSmbdoPhoneType = {
  BUSINESS_PHONE: 'BUSINESS_PHONE',
  MOBILE_PHONE: 'MOBILE_PHONE',
  ALTERNATE_PHONE: 'ALTERNATE_PHONE',
} as const;

/**
 * Phone Number Information of the account

 */
export interface PhoneSmbdo {
  countryCode: string;
  phoneNumber: string;
  phoneType: PhoneSmbdoPhoneType;
}

/**
 * Details of an organization.
 */
export interface OrganizationDetails {
  /**
   * @minItems 1
   * @maxItems 5
   */
  addresses?: AddressDto[];
  /**
   * @minItems 0
   * @maxItems 100
   */
  associatedCountries?: CountryCodeIsoAlpha2[];
  countryOfFormation?: CountryCodeIsoAlpha2;
  dbaName?: DbaName;
  entitiesInOwnership?: EntitiesInOwnership;
  industry?: OrganizationIndustry;
  industryCategory?: OrganizationIndustryCategory;
  industryType?: OrganizationIndustryType;
  jurisdiction?: CountryCodeIsoAlpha2;
  mcc?: MerchantCategoryCode;
  organizationDescription?: OrganizationDescription;
  /**
   * @minItems 0
   * @maxItems 6
   */
  organizationIds?: OrganizationIdentityDto[];
  organizationName?: OrganizationName;
  organizationType?: OrganizationType;
  phone?: PhoneSmbdo;
  /**
   * The list of additional merchant category codes describing  industries that the business is in.

   * @minItems 0
   * @maxItems 10
   */
  secondaryMccList?: MerchantCategoryCode[];
  socialMedia?: SocialMediaList;
  website?: Website;
  websiteAvailable?: WebsiteAvailable;
  yearOfFormation?: YearOfFormation;
}

/**
 * If `jobTitle` is Other, then job title description is required.

 * @minLength 0
 * @maxLength 50
 */
export type IndividualJobTitleDescription = string;

/**
 * Job title in the case of party type being an `INDIVIDUAL`. Job title is a required field for Controllers. Also, If Privately Owned Business is selected as the business type, Job Title should be a required field for Decision Makers. e.g. CEO|CFO|COO|President|Chairman|Senior Branch Manager|Other

 * @minLength 0
 */
export type IndividualJobTitle = string;

/**
 * Details of an individual.
 */
export interface IndividualDetails {
  /**
   * @minItems 1
   * @maxItems 5
   */
  addresses?: AddressDto[];
  birthDate?: BirthDate;
  countryOfResidence?: CountryCodeIsoAlpha2;
  firstName?: FirstName;
  /**
   * @minItems 0
   * @maxItems 16
   */
  individualIds?: IndividualIdentity[];
  jobTitle?: IndividualJobTitle;
  jobTitleDescription?: IndividualJobTitleDescription;
  lastName?: LastName;
  middleName?: MiddleName;
  nameSuffix?: NameSuffix;
  natureOfOwnership?: NatureOfOwnership;
  phone?: PhoneSmbdo;
  socialMedia?: SocialMediaList;
  soleOwner?: SoleOwner;
}

/**
 * `idType` denotes the type of taxpayer identification number (e.g. Social Security Number or Individual Taxpayer Identification Number). A Social Security Number or Individual Taxpayer Identification Number is accepted for an owner or controller individual. Decision makers do not require any tax identifier.

 */
export type IndividualIdentityIdType =
  (typeof IndividualIdentityIdType)[keyof typeof IndividualIdentityIdType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const IndividualIdentityIdType = {
  SSN: 'SSN',
  ITIN: 'ITIN',
  NATIONAL_ID: 'NATIONAL_ID',
  DRIVERS_LICENSE: 'DRIVERS_LICENSE',
  PASSPORT: 'PASSPORT',
  SOCIAL_INSURANCE_NUMBER: 'SOCIAL_INSURANCE_NUMBER',
  OTHER_GOVERNMENT_ID: 'OTHER_GOVERNMENT_ID',
} as const;

/**
 * An individual's identification.
 */
export interface IndividualIdentity {
  /** Description of the ID. */
  description?: string;
  expiryDate?: string;
  /** `idType` denotes the type of taxpayer identification number (e.g. Social Security Number or Individual Taxpayer Identification Number). A Social Security Number or Individual Taxpayer Identification Number is accepted for an owner or controller individual. Decision makers do not require any tax identifier.
   */
  idType: IndividualIdentityIdType;
  /** Identification issuer country code e.g. US */
  issuer: string;
  /**
   * Value of the identification type. EIN/SSN/ITIN must be of 9 digits.
   * @minLength 1
   * @maxLength 20
   */
  value: string;
}

/**
 * The name suffix used by the party
 * @minLength 1
 * @maxLength 5
 */
export type NameSuffix = string;

/**
 * Last name of the individual in case of party type being an individual like Owners, Controllers and Decision Makers.

 * @minLength 2
 * @maxLength 30
 */
export type LastName = string;

/**
 * Middle name of the individual in case of party type being an individual.

 * @minLength 0
 * @maxLength 30
 */
export type MiddleName = string;

/**
 * First name of the individual in case of party type being an individual like Owners, Controllers and Decision Makers.

 * @minLength 2
 * @maxLength 30
 */
export type FirstName = string;

/**
 * Country code in ISO alpha-2 format.
 * @minLength 2
 * @maxLength 2
 */
export type CountryCodeIsoAlpha2 = string;

/**
 * The date of birth (yyyy-MM-dd) of the individual in case of party type being an individual.

 */
export type BirthDate = string;

/**
 * Type of address.
 */
export type AddressDtoAddressType =
  (typeof AddressDtoAddressType)[keyof typeof AddressDtoAddressType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AddressDtoAddressType = {
  LEGAL_ADDRESS: 'LEGAL_ADDRESS',
  MAILING_ADDRESS: 'MAILING_ADDRESS',
  BUSINESS_ADDRESS: 'BUSINESS_ADDRESS',
  RESIDENTIAL_ADDRESS: 'RESIDENTIAL_ADDRESS',
} as const;

/**
 * An address.
 */
export interface AddressDto {
  /**
   * The first line must not be a PO Box and must begin with a number. Each line has a maximum of 60 characters.
   * @minItems 1
   * @maxItems 5
   */
  addressLines: string[];
  /** Type of address. */
  addressType?: AddressDtoAddressType;
  /**
   * city has a maximum of 30 characters.
   * @maxLength 34
   */
  city: string;
  /** Country code in alpha-2 format. */
  country: string;
  /**
   * Postal/ZIP code.
   * @maxLength 10
   */
  postalCode: string;
  /**
   * State code in alpha-2 format. State is mandatory for countries like United States. Refer to subdivision codes in [ISO-3166-2](https://www.iso.org/obp/ui/#search/code/).
   * @minLength 1
   * @maxLength 3
   */
  state?: string;
}